cc: rename encryptor() -> get_encryptor()
