+++ monkey/monkey_island/cc/server_utils/encryptor.py
@@ -62,5 +62,5 @@ def initialize_encryptor(data_dir):
     _encryptor = Encryptor(data_dir)
 
 
-def encryptor():
+def get_encryptor():
     return _encryptor
