+++ monkey/monkey_island/cc/services/config.py
@@ -8,7 +8,7 @@ from jsonschema import Draft4Validator, validators
 import monkey_island.cc.environment.environment_singleton as env_singleton
 import monkey_island.cc.services.post_breach_files
 from monkey_island.cc.database import mongo
-from monkey_island.cc.server_utils.encryptor import encryptor
+from monkey_island.cc.server_utils.encryptor import get_encryptor
 from monkey_island.cc.services.utils.network_utils import local_ip_addresses
 from monkey_island.cc.services.config_schema.config_schema import SCHEMA
 
@@ -75,9 +75,9 @@ class ConfigService:
         if should_decrypt:
             if config_key_as_arr in ENCRYPTED_CONFIG_VALUES:
                 if isinstance(config, str):
-                    config = encryptor().dec(config)
+                    config = get_encryptor().dec(config)
                 elif isinstance(config, list):
-                    config = [encryptor().dec(x) for x in config]
+                    config = [get_encryptor().dec(x) for x in config]
         return config
 
     @staticmethod
@@ -112,7 +112,7 @@ class ConfigService:
         if item_value in items_from_config:
             return
         if should_encrypt:
-            item_value = encryptor().enc(item_value)
+            item_value = get_encryptor().enc(item_value)
         mongo.db.config.update(
             {'name': 'newconfig'},
             {'$addToSet': {item_key: item_value}},
@@ -297,9 +297,9 @@ class ConfigService:
                 if flat_config[key] and isinstance(flat_config[key][0], dict) and 'public_key' in flat_config[key][0]:
                     flat_config[key] = [ConfigService.decrypt_ssh_key_pair(item) for item in flat_config[key]]
                 else:
-                    flat_config[key] = [encryptor().dec(item) for item in flat_config[key]]
+                    flat_config[key] = [get_encryptor().dec(item) for item in flat_config[key]]
             else:
-                flat_config[key] = encryptor().dec(flat_config[key])
+                flat_config[key] = get_encryptor().dec(flat_config[key])
         return flat_config
 
     @staticmethod
@@ -320,19 +320,19 @@ class ConfigService:
                         config_arr[i] = ConfigService.decrypt_ssh_key_pair(config_arr[i]) if is_decrypt else \
                             ConfigService.decrypt_ssh_key_pair(config_arr[i], True)
                     else:
-                        config_arr[i] = encryptor().dec(config_arr[i]) if is_decrypt else encryptor().enc(config_arr[i])
+                        config_arr[i] = get_encryptor().dec(config_arr[i]) if is_decrypt else get_encryptor().enc(config_arr[i])
             else:
                 parent_config_arr[config_arr_as_array[-1]] = \
-                    encryptor().dec(config_arr) if is_decrypt else encryptor().enc(config_arr)
+                    get_encryptor().dec(config_arr) if is_decrypt else get_encryptor().enc(config_arr)
 
     @staticmethod
     def decrypt_ssh_key_pair(pair, encrypt=False):
         if encrypt:
-            pair['public_key'] = encryptor().enc(pair['public_key'])
-            pair['private_key'] = encryptor().enc(pair['private_key'])
+            pair['public_key'] = get_encryptor().enc(pair['public_key'])
+            pair['private_key'] = get_encryptor().enc(pair['private_key'])
         else:
-            pair['public_key'] = encryptor().dec(pair['public_key'])
-            pair['private_key'] = encryptor().dec(pair['private_key'])
+            pair['public_key'] = get_encryptor().dec(pair['public_key'])
+            pair['private_key'] = get_encryptor().dec(pair['private_key'])
         return pair
 
     @staticmethod
