+++ monkey/monkey_island/cc/test_encryptor.py
@@ -1,7 +1,7 @@
 import os
 
 from monkey_island.cc.server_utils.consts import MONKEY_ISLAND_ABS_PATH
-from monkey_island.cc.server_utils.encryptor import initialize_encryptor, encryptor
+from monkey_island.cc.server_utils.encryptor import initialize_encryptor, get_encryptor
 
 
 TEST_DATA_DIR = os.path.join(MONKEY_ISLAND_ABS_PATH, "cc", "testing")
@@ -14,13 +14,13 @@ CYPHERTEXT = "vKgvD6SjRyIh1dh2AM/rnTa0NI/vjfwnbZLbMocWtE4e42WJmSUz2ordtbQrH1Fq"
 def test_aes_cbc_encryption():
     initialize_encryptor(TEST_DATA_DIR)
 
-    assert encryptor().enc(PLAINTEXT) != PLAINTEXT
+    assert get_encryptor().enc(PLAINTEXT) != PLAINTEXT
 
 
 def test_aes_cbc_decryption():
     initialize_encryptor(TEST_DATA_DIR)
 
-    assert encryptor().dec(CYPHERTEXT) == PLAINTEXT
+    assert get_encryptor().dec(CYPHERTEXT) == PLAINTEXT
 
 
 def test_create_new_password_file(tmpdir):
